Synchronize panel visibility between toolbars
Added a call to `EquationToolBar.TryHideVisiblePanel()` in the
`SetActivePanel` method of the `CharacterToolBar` class to hide
any visible panel in `EquationToolBar` when activating a panel
in `CharacterToolBar`.

Similarly, added a call to `CharacterToolBar.TryHideVisiblePanel()`
in the `SetActivePanel` method of the `EquationToolBar` class to
hide any visible panel in `CharacterToolBar` when activating a
panel in `EquationToolBar`.